backend/atomic/email/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-restx==1.3.0
requests==2.32.3
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pika
